requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
jinja2==3.1.6
sqlalchemy==2.0.44
psycopg[binary]==3.2.13
alembic==1.16.5
pydantic-settings==2.8.1
python-multipart==0.0.20
email-validator==2.2.0
pytest==8.3.5
httpx==0.28.1
